+++ tailwind.config.js
@@ -6,7 +6,6 @@ module.exports = {
   },
   variants: {
     extend: {
-      textColor: ['visited'],
       borderWidth: ['hover']
     },
   },
